requirements.txt
load-dotenv==0.1.0
python-dotenv==0.21.1
python-telegram-bot==13.15
requests==2.28.2
